Update static code analyzer and general code clean up
* Imagine: Add support for clang-tidy to build system and remove cppcheck
* Imagine: Update C++ language standard to C++23
* Add a stale PR/issue workflow job

+++ EmuFramework/src/EmuInput.cc
@@ -432,7 +432,7 @@ void genericMultiplayerTranspose(KeyConfig::KeyArray &key, unsigned player, unsi
 {
 	iterateTimes(EmuSystem::maxPlayers, i)
 	{
-		if(player && i == player)
+		if(player && (unsigned)i == player)
 		{
 			//logMsg("moving to player %d map", i);
 			std::copy_n(&key[category[startCategory].configOffset], category[startCategory].keys, &key[category[i+startCategory].configOffset]);
